+++ changelog.md
@@ -12,6 +12,7 @@
 ### Improvements
 - Games show as columns on wide screens to save space
 - Removed confirmation dialogue when adding unregistered players to stages
+- New card in Faction settings to view the faction in the Forge if linked
 
 ### Bug Fixes
 - Expansion image bug where the image could be unintentionally deleted